Fix version fetcher
